# frontend/openapi.json
{
  "components": {
    "schemas": {
      "AskRequest": {
        "properties": {
          "message": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "AskRequest",
        "type": "object"
      },
      "AskResponse": {
        "additionalProperties": true,
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "confidence": {
            "title": "Confidence",
            "type": "string"
          },
          "intent": {
            "title": "Intent",
            "type": "string"
          },
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/AskSource"
            },
            "title": "Sources",
            "type": "array"
          }
        },
        "required": [
          "mode",
          "intent",
          "confidence",
          "answer",
          "sources",
          "note"
        ],
        "title": "AskResponse",
        "type": "object"
      },
      "AskSource": {
        "additionalProperties": true,
        "properties": {
          "english": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "English"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "verification_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verification Status"
          },
          "verified": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified"
          },
          "zomi": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zomi"
          }
        },
        "required": [
          "type"
        ],
        "title": "AskSource",
        "type": "object"
      },
...
      "BibleChapterResponse": {
        "additionalProperties": true,
        "properties": {
          "book_english": {
            "title": "Book English",
            "type": "string"
          },
          "book_id": {
            "title": "Book Id",
            "type": "integer"
          },
          "book_zomi": {
            "title": "Book Zomi",
            "type": "string"
          },
          "chapter": {
            "title": "Chapter",
            "type": "integer"
          },
          "verse_count": {
            "title": "Verse Count",
            "type": "integer"
          },
          "verses": {
            "items": {
              "$ref": "#/components/schemas/BibleVerse"
            },
            "title": "Verses",
            "type": "array"
          }
        },
        "required": [
          "book_id",
          "chapter",
          "book_english",
          "book_zomi",
          "verse_count",
          "verses"
        ],
        "title": "BibleChapterResponse",
        "type": "object"
      },
      "BibleSearchResponse": {
        "additionalProperties": true,
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/BibleVerse"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "query",
          "count",
          "results"
        ],
        "title": "BibleSearchResponse",
        "type": "object"
      },
      "BibleVerse": {
        "additionalProperties": true,
        "properties": {
          "book_english": {
            "title": "Book English",
            "type": "string"
          },
          "book_id": {
            "title": "Book Id",
            "type": "integer"
          },
          "book_zomi": {
            "title": "Book Zomi",
            "type": "string"
          },
          "changes_applied": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Changes Applied"
          },
          "chapter": {
            "title": "Chapter",
            "type": "integer"
          },
          "english": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "English"
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          },
          "testament": {
            "title": "Testament",
            "type": "string"
          },
          "verse": {
            "title": "Verse",
            "type": "integer"
          },
          "zomi_iso": {
            "title": "Zomi Iso",
            "type": "string"
          },
          "zomi_original": {
            "title": "Zomi Original",
            "type": "string"
          }
        },
        "required": [
          "testament",
          "book_id",
          "book_english",
          "book_zomi",
          "chapter",
          "verse",
          "reference",
          "zomi_original",
          "zomi_iso"
        ],
        "title": "BibleVerse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IdentityResponse": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "locale": {
            "title": "Locale",
            "type": "string"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "locale",
          "roles"
        ],
        "title": "IdentityResponse",
        "type": "object"
      },
...
      "RoleAssignmentResponse": {
        "properties": {
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "roles"
        ],
        "title": "RoleAssignmentResponse",
        "type": "object"
      },
      "RoleName": {
        "enum": [
          "learner",
          "contributor",
          "reviewer",
          "editor",
          "moderator",
          "administrator"
        ],
        "title": "RoleName",
        "type": "string"
      },
      "TranslateMatch": {
        "additionalProperties": true,
        "properties": {
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain"
          },
          "english": {
            "title": "English",
            "type": "string"
          },
          "exact": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exact"
          },
          "flagged": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flagged"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "part_of_speech": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Of Speech"
          },
          "quality_flags": {
            "items": {
              "type": "string"
            },
            "title": "Quality Flags",
            "type": "array"
          },
          "quality_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Score"
          },
          "reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "verification_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verification Status"
          },
          "verified": {
            "default": false,
            "title": "Verified",
            "type": "boolean"
          },
          "zomi": {
            "title": "Zomi",
            "type": "string"
          }
        },
        "required": [
          "source",
          "english",
          "zomi"
        ],
        "title": "TranslateMatch",
        "type": "object"
      },
...
      "UnifiedSearchResponse": {
        "additionalProperties": true,
        "properties": {
          "bible": {
            "items": {
              "$ref": "#/components/schemas/BibleVerse"
            },
            "title": "Bible",
            "type": "array"
          },
          "dictionary": {
            "items": {
              "$ref": "#/components/schemas/DictionaryEntry"
            },
            "title": "Dictionary",
            "type": "array"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "translations": {
            "items": {
              "$ref": "#/components/schemas/TranslateMatch"
            },
            "title": "Translations",
            "type": "array"
          }
        },
        "required": [
          "query",
          "dictionary",
          "translations",
          "bible"
        ],
        "title": "UnifiedSearchResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Zomi language platform — dictionary, translation retrieval, Bible, and learning.",
    "title": "Siamsil API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/ai/ask": {
      "post": {
        "operationId": "ask_api_ai_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ask",
        "tags": [
          "ai"
        ]
      }
    },
...
    "/api/bible/chapter": {
      "get": {
        "operationId": "get_chapter_api_bible_chapter_get",
        "parameters": [
          {
            "in": "query",
            "name": "book_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Book Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "chapter",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Chapter",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BibleChapterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Chapter",
        "tags": [
          "bible"
        ]
      }
    },
...
    "/api/bible/search": {
      "get": {
        "operationId": "search_bible_api_bible_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BibleSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Bible",
        "tags": [
          "bible"
        ]
      }
    },
...
    "/api/search": {
      "get": {
        "operationId": "unified_search_api_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 8,
              "maximum": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unified Search",
        "tags": [
          "search"
        ]
      }
    },
...
    "/api/v1/ai/ask": {
      "post": {
        "operationId": "ask_api_v1_ai_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ask",
        "tags": [
          "ai"
        ]
      }
    },
...
    "/api/v1/bible/chapter": {
      "get": {
        "operationId": "get_chapter_api_v1_bible_chapter_get",
        "parameters": [
          {
            "in": "query",
            "name": "book_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Book Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "chapter",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Chapter",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BibleChapterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Chapter",
        "tags": [
          "bible"
        ]
      }
    },
...
    "/api/v1/bible/search": {
      "get": {
        "operationId": "search_bible_api_v1_bible_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BibleSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Bible",
        "tags": [
          "bible"
        ]
      }
    },
...
    "/api/v1/identity/me": {
      "get": {
        "operationId": "current_identity_api_v1_identity_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Current Identity",
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/identity/users/{user_id}/roles/{role}": {
      "post": {
        "operationId": "assign_role_api_v1_identity_users__user_id__roles__role__post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "role",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RoleName"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleAssignmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Assign Role",
        "tags": [
          "identity"
        ]
      }
    },
...
    "/api/v1/search": {
      "get": {
        "operationId": "unified_search_api_v1_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 8,
              "maximum": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unified Search",
        "tags": [
          "search"
        ]
      }
    },
...
  }
}